+++ .gitignore
@@ -7,4 +7,4 @@ Thumbs.db
 *.psd
 *.ai
 *.pdf
-build/js/main.js
+build/js/
