+++ src/main.js
@@ -1,5 +1,10 @@
+/**
+ * @fileoverview Список отелей: загрузка, включение фильтрации и постраничная
+ * отрисовка)
+ */
+
 'use strict';
 
 require('./form.js');
 require('./game.js');
-require('./reviews.js');
+require('./reviews-block/reviews.js');
