+++ src/reviews-block/filter/filter-type.js
@@ -1,0 +1,14 @@
+/**
+ * @fileoverview List of available filters
+ */
+
+'use strict';
+
+/** @enum {number} */
+module.exports = {
+  'ALL': 'reviews-all',
+  'RECENT': 'reviews-recent',
+  'GOOD': 'reviews-good',
+  'BAD': 'reviews-bad',
+  'POPULAR': 'reviews-popular'
+};
